+++ libs/ts/classes/classes-domain/src/index.ts
@@ -4,8 +4,8 @@ export * from './lib/models/discount/discount';
 export * from './lib/models/discount/amount';
 export * from './lib/models/discount/discount.dbo';
 export * from './lib/models/discount/discount-evaluator';
-export * from './lib/models/discount/classes-discount';
-export * from './lib/models/discount/basket-discount';
+export * from './lib/models/discount/classes-discount.abstract';
+export * from './lib/models/discount/basket-discount.abstract';
 export * from './lib/models/semester-enrollment';
 export * from './lib/models/path/path-db-entry';
 export * from './lib/models/path/unit-db-entry';
