+++ libs/ts/classes/classes-domain/src/lib/models/discount/basket-discount.abstract.ts
