+++ libs/ts/classes/classes-domain/src/lib/models/discount/classes-discount.abstract.ts
